refactor(shared): promote query feature to @langfuse/shared (LFE-9806)
