+++ packages/shared/src/env.ts
@@ -99,6 +99,20 @@ const EnvSchema = z.object({
   CLICKHOUSE_UPDATE_PARALLEL_MODE: z
     .enum(["sync", "async", "auto"])
     .default("auto"),
+  CLICKHOUSE_MAX_BYTES_BEFORE_EXTERNAL_GROUP_BY: z.coerce
+    .number()
+    .default(32_000_000_000), // ~32GB
+  CLICKHOUSE_USE_QUERY_CONDITION_CACHE: z
+    .enum(["true", "false"])
+    .default("false"),
+  LANGFUSE_ENABLE_SINGLE_LEVEL_QUERY_OPTIMIZATION: z
+    .enum(["true", "false"])
+    .default("false"),
+  LANGFUSE_ROOT_EVENT_CONDITION_MAX_WINDOW_HOURS: z.coerce
+    .number()
+    .int()
+    .nonnegative()
+    .default(168), // 7 days
 
   LANGFUSE_INGESTION_QUEUE_DELAY_MS: z.coerce
     .number()
