+++ packages/shared/src/features/query/dashboardUiTableToViewMapping.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
-import { singleFilter } from "@langfuse/shared";
-import { type views } from "@/src/features/query/types";
+import { singleFilter } from "../../interfaces/filters";
+import { type views } from "./types";
 
 // Exported to silence @typescript-eslint/no-unused-vars v8 warning
 // (used for type extraction via typeof, which is a legitimate pattern)
