+++ packages/shared/src/features/query/index.ts
@@ -1,3 +1,4 @@
 export * from "./dataModel";
 export * from "./dashboardUiTableToViewMapping";
 export * from "./types";
+export * from "./validateQuery";
