+++ packages/shared/src/features/query/server/nullIfEmptyFilter.ts
