+++ packages/shared/src/features/query/server/queryBuilder.ts
@@ -1,8 +1,12 @@
+import { randomUUID } from "crypto";
 import { type z } from "zod";
 import {
   convertDateToClickhouseDateTime,
   shouldSkipObservationsFinal,
-} from "@langfuse/shared/src/server";
+  FilterList,
+  createFilterFromFilterState,
+  type Filter,
+} from "../../../server";
 import type {
   QueryType,
   ViewDeclarationType,
@@ -15,14 +19,9 @@ import {
   query as queryModel,
   getValidAggregationsForMeasureType,
 } from "../types";
-import { getViewDeclaration } from "@/src/features/query/dataModel";
-import {
-  FilterList,
-  createFilterFromFilterState,
-  type Filter,
-} from "@langfuse/shared/src/server";
-import { InvalidRequestError } from "@langfuse/shared";
-import { env } from "@/src/env.mjs";
+import { getViewDeclaration } from "../dataModel";
+import { InvalidRequestError } from "../../../errors";
+import { env } from "../../../env";
 import { NULL_IF_EMPTY_RE } from "./nullIfEmptyFilter";
 
 type AppliedDimensionType = {
@@ -86,18 +85,20 @@ export class QueryBuilder {
         return `quantile(0.95)(${metric.alias || metric.sql})`;
       case "p99":
         return `quantile(0.99)(${metric.alias || metric.sql})`;
-      case "histogram":
+      case "histogram": {
         // Get histogram bins from chart config, fallback to 10
         const bins = this.chartConfig?.bins ?? 10;
         return `histogram(${bins})(toFloat64(${metric.alias || metric.sql}))`;
+      }
       case "uniq":
         return `uniq(${metric.alias || metric.sql})`;
-      default:
+      default: {
         // eslint-disable-next-line @typescript-eslint/no-unused-vars
         const exhaustiveCheck: never = metric.aggregation;
         throw new InvalidRequestError(
           `Invalid aggregation: ${metric.aggregation}`,
         );
+      }
     }
   }
 
@@ -717,12 +718,13 @@ export class QueryBuilder {
         throw new Error(
           `Granularity 'auto' is not supported for getTimeDimensionSql`,
         );
-      default:
+      default: {
         // eslint-disable-next-line @typescript-eslint/no-unused-vars
         const exhaustiveCheck: never = granularity;
         throw new InvalidRequestError(
           `Invalid time granularity: ${granularity}. Must be one of minute, hour, day, week, month`,
         );
+      }
     }
   }
 
@@ -1356,7 +1358,7 @@ export class QueryBuilder {
       if (thresholdHours === 0 || windowHours <= thresholdHours) {
         // Falls back gracefully: if no root events exist in the window at all
         // (e.g. parent_span_id is never populated), the filter is skipped via NOT EXISTS.
-        const uid = crypto.randomUUID().replace(/-/g, "").slice(0, 8);
+        const uid = randomUUID().replace(/-/g, "").slice(0, 8);
         const fromP = `subFrom${uid}`;
         const toP = `subTo${uid}`;
         const projP = `subProj${uid}`;
