+++ packages/shared/src/index.ts
@@ -99,3 +99,9 @@ export * from "./utils/IORepresentation";
 
 // analytics integrations (client-safe)
 export * from "./features/analytics-integrations";
+
+// query (dashboard / monitor data model)
+export * from "./features/query/types";
+export * from "./features/query/dataModel";
+export * from "./features/query/dashboardUiTableToViewMapping";
+export * from "./features/query/validateQuery";
