+++ web/src/__tests__/server/dashboard-widget-version.servertest.ts
@@ -6,11 +6,11 @@ import { prisma } from "@langfuse/shared/src/db";
 import { appRouter } from "@/src/server/api/root";
 import { createInnerTRPCContext } from "@/src/server/api/trpc";
 import type { Session } from "next-auth";
-import { requiresV2 } from "@/src/features/query/dataModel";
+import { requiresV2 } from "@langfuse/shared";
 import {
   mapLegacyUiTableFilterToView,
   mapWidgetUiTableFilterToView,
-} from "@/src/features/query";
+} from "@langfuse/shared";
 
 describe("dashboard widget minVersion", () => {
   let projectId: string;
