+++ web/src/__tests__/server/metrics-api.servertest.ts
@@ -6,7 +6,7 @@ import {
 } from "@/src/__tests__/test-utils";
 import { GetMetricsV1Response } from "@/src/features/public-api/types/metrics";
 import { createBasicAuthHeader } from "@langfuse/shared/src/server";
-import { type QueryType } from "@/src/features/query/types";
+import { type QueryType } from "@langfuse/shared";
 import {
   createTrace,
   createObservation,
