+++ web/src/__tests__/server/nullIfEmptyFilter.servertest.ts
@@ -1,4 +1,4 @@
-import { NULL_IF_EMPTY_RE } from "@/src/features/query/server/nullIfEmptyFilter";
+import { NULL_IF_EMPTY_RE } from "@langfuse/shared/src/server";
 import {
   StringFilter,
   StringOptionsFilter,
