+++ web/src/__tests__/server/queryBuilderDashboards.servertest.ts
@@ -5,9 +5,9 @@ import {
   createObservation,
   createObservationsCh,
 } from "@langfuse/shared/src/server";
-import { type QueryType } from "@/src/features/query/types";
-import { executeQuery } from "@/src/features/query/server/queryExecutor";
-import { mapLegacyUiTableFilterToView } from "@/src/features/query";
+import { type QueryType } from "@langfuse/shared";
+import { executeQuery } from "@langfuse/shared/src/server";
+import { mapLegacyUiTableFilterToView } from "@langfuse/shared";
 
 describe("selfServeDashboards", () => {
   // Single project ID for all tests
