+++ web/src/__tests__/server/queryBuilderSQLI.servertest.ts
@@ -1,7 +1,7 @@
 import { randomUUID } from "crypto";
-import { QueryBuilder } from "@/src/features/query/server/queryBuilder";
-import { type QueryType } from "@/src/features/query/types";
-import { executeQuery } from "@/src/features/query/server/queryExecutor";
+import { QueryBuilder } from "@langfuse/shared/src/server";
+import { type QueryType } from "@langfuse/shared";
+import { executeQuery } from "@langfuse/shared/src/server";
 import { InvalidRequestError } from "@langfuse/shared";
 
 /**
