+++ web/src/features/dashboard/components/ScoresTable.tsx
@@ -5,7 +5,7 @@ import {
   type ScoreSourceType,
   type FilterState,
 } from "@langfuse/shared";
-import { type ViewVersion } from "@/src/features/query";
+import { type ViewVersion } from "@langfuse/shared";
 import { api } from "@/src/utils/api";
 import { compactNumberFormatter } from "@/src/utils/numbers";
 import { RightAlignedCell } from "./RightAlignedCell";
