+++ web/src/pages/project/[projectId]/widgets/[widgetId]/index.tsx
@@ -5,7 +5,7 @@ import { WidgetForm } from "@/src/features/widgets/components/WidgetForm";
 import { showErrorToast } from "@/src/features/notifications/showErrorToast";
 import { showSuccessToast } from "@/src/features/notifications/showSuccessToast";
 import { type DashboardWidgetChartType } from "@langfuse/shared/src/db";
-import { type views, type metricAggregations } from "@/src/features/query";
+import { type views, type metricAggregations } from "@langfuse/shared";
 import { type z } from "zod";
 import { type WidgetChartConfig } from "@/src/features/widgets/utils";
